e3: Refactor to monolithic train() function
Removing the original step() and train() methods
from StoppableTrainer, renaming run() to train().

This change might look stupid, but it enables access
to many more important training-related variables
in the training shell. The previous design was more
modular, but it made the shell useless.

I will try to find a way to re-modularize it later, but for
now, I need to work on more important bugs where I need the shell.