Fix segfaults
The problem was that sometimes (but not always)
a segfault happens if the model runs in inference mode
when the input data shape is different than the one it was
trained with. Both smaller and larger inputs
can lead to this. I don't know why and how this can be fixed,
but for now, I will just use the training size for test inference.
I guess this is a library bug, maybe fixed in a newer release.

I will clean up the mess later...