models: Add 3d U-Net model
This is quick and dirty 3d adaptation (mostly copy-pasted) of the
(2d) U-Net implementation by Jackson Huang
(https://github.com/jaxony/unet-pytorch), which is based on
https://arxiv.org/abs/1505.04597 with some additional improvements
(see unet.UNet class docstring).

This is not meant as an implementation of the "3D U-Net" proposed in
https://arxiv.org/abs/1606.06650, but of course there are many similarities.

Training with "$ python scripts/train.py unet" works, but it is far from
perfect (see TODO comments).