slice_h5: Retry on HDF5 read error
Testing with 6d8561e has revealed that the intermittent
read errors usually don't occur again if a read attempt
at exactly the same coordinates in the same file is repeated.

For the (apparently very unlikely) case that the next attempt
also fails, I have set max_retries to 5 for each batch.
If it fails more than 5 times in a row, the current coordinates
are skipped.

Since the errors are not deterministic, I guess they are either caused
by a concurrency issue in PyTorch's DataLoader, in HDF5/h5py or maybe it's
even a filesystem issue.
(One of the error messages can be found in the commit message of e1a55ed.)