PatchCreator: Workaround for invalid target check
The previous check assumed that self.c_target was the number of unique target
classes in the data set (which is wrong - it actually counts the number of
channels of the target tensors).
Since we don't know this number, I had to resort to a dumb workaround.
(We could support setting the number of classes via constructor args,
but I hope we can fix #10 soon instead).