examples: Add unet train script, remove train.py
This example is a reduced and cleaned up version of the former
scripts/train.py.
I deleted most options, comments and conditionals that are not
absolutely necessary for it to work, so the code is more readable.
Almost everything is hardcoded now, but can be easily changed. It should
be seen as a template for own training scripts.
More examples to follow.

The old scripts/train.py can still be found at
https://github.com/ELEKTRONN/elektronn3/blob/a18301d19e/scripts/train.py
if anyone needs it. It could still be useful as a reference.

Closes #4 (see option 2.).