data: Start work on transformation pipeline API
WIP. Not sure about the API design of transforms.py. I decided to
use the general transform specification

    transformed_inp, transformed_target = transform(inp, target)

This has the disadvantage that it won't work with transforms that
are written in the current torchvsion style

    transformed_inp = transform(inp)

but we will definitely need joint input-target-transforms, so I
thought it would be best to use a generalized form that *can* work with
targets, but can also ignore them (just returning the unmodified
target).

The most obvious alternative would have been to use two kinds of
transforms: One that only works on the input (like torchvision's) and
one that does joint transforms to inputs and targets.
I decided against that because it would be more confusing to have two
kinds of incompatible transforms that have to be clearly distinguishable
for the user and would also require at least one more parameter in dataset
constructors (transforms=..., joint_transforms=..., maybe even
target_transforms=...).

Related discussion: https://github.com/pytorch/vision/issues/230
There are some good points for other API designs in there, so we might
reconsider this later.