Move std and mean calculation out of PatchCreator
Also remove the respective properties and attributes

Dataset statistics only have to be
computed once per dataset, so I think it's okay to not do this
automatically anymore, especially since the algorithms used for
computing them are not well thought out.

Helpers for calculating means and stds are now in the data.utils module.