data: Uncouple transform matrix calc from slicing
Refactoring to simplify code and clean up the mixing of
slicing/disk IO and the (purely mathematical) calculation of
homogeneous coordinate transformation matrices.

Includes a few renames and documentation updates where they were
necessary.

I also removed "strip_2d" stuff because it not used and untested.
This kind of 2D functionality can return if there is demand for it,
after the big coord-transform/slicing code refactoring is complete.