previews: Uncouple preview pred from PatchCreator
Preview batches can now be loaded from arbitrary locations and no longer
necessarily depend on validation data.
Targets are no longer plotted with previews, because the user is
expected to know the GT targets for the selected region. By not
requiring targets, we are now free to choose unlabeled data for
previews.