warping: Fix out-of-bounds read in numba gufunc
The upper bound for slice indices was off by one.
This was causing out-of-bounds reads that were not detected
because the numba-jitted gufuncs don't do any bounds checking and
silently read from unallocated memory regions if indices are not within
the bounds of the source array.

Also: Add a docstring with warnings about map_coordinates_nearest()
(TODO: map_coordinates_linear() should receive such a docstring, too!)

Resolves #10