warping: Fix out-of-bound read (again), add checks
a60904e5 wasn't actually the correct fix for the out-of-bounds read
issue. The calculation of hi and hi_targ is now reverted to the old
state. The actual fix was to narrow down both OOB checks by 1 in all
axes (see lines 352 and 356)

Also add some documentation and some temporary checks that will be
left in for a while to make sure we've actually fixed the issue now.