visualization: Support smaller outputs and targets
If the neural network output and targets are smaller than inputs (e.g.
when using the new conv_mode='valid' in UNet), outputs and targets are
now zero-padded for visualization in tensorboard to make the
corresponding regions more clear.

This change works for training and validation samples, but is not yet
effective for preview predictions. There are further changes required in
the inference module for the latter to work.