example: Switch to planar_blocks=(0,)
Only the highest-resolution blocks (in the encoder and decoder pathways)
of the 3D U-Net now effectively work in 2D and the image depth is only
considered in deeper layers, by using actual 3D convolutions.
This is now feasible thanks to the new layers introduced in 18cab5b.
The previous setting planar_blocks=(1,) was just used as a workaround,
because it didn't suffer as much from the performance issues described
in 18cab5b.

First test trainings show comparable training and validation metric scores to
the previous setting.