example: Support alternative dataset, refactor
If the envvar CLUSTER is set to WHOLEBRAIN, the example training script
now tries to load a larger, but private data set. This will probably be
removed some day, but for now this change makes rapid experimentation
less troublesome for wholebrain compute cluster users.
The training example is now also easier to adapt to other data sets.