inference: Support auto-padding in-/out shapes
This relaxes the constraint that out_shape has to be divisible by
tile_shape and thereby makes finding valid configurations for
Predictor objects much easier. When shapes don't match and strict_shapes
is set to False, automatic padding is used to make them compatible, but
at the price of significant wasted computations in the padded regions.
The user will be informed about automatic padding in the logger output.
If shapes do match, everything behaves as before.

Limitations:
- Requires that the spatial output shape is the same as the spatial
  input shape.
- The info log message about wasted compute doesn't account for
  overlap_shape and is therefore much lower if overlap_shape is
  non-zero.

Closes #29, because it is a generalization of this PR.

Co-authored-by: Norbert Pfeiler <norbert.pfeiler+git@gmail.com>