example: Use CyclicLR and re-enable combined loss
The divergence issues with CrossEntropyLoss in recent versions of the
train_unet_neurodata example were caused by the sudden learning rate
spikes of the CosineAnnealingWarmRestarts scheduler (even with the
maximum LR set to an order of magnitude lower, this was still the case).
These spikes are not an issue for tranings with pure DiceLoss, but since
we want to support CrossEntropy, we are now switching to a Cyclical
Learning Rates scheduler, which solves this issue by gradually warming
up the training with increasing learning rates instead of directly
jumping from the minimum LR to the maximum LR.

The new base_lr and max_lr hyperparams were determined with the
learning rate range test described in https://arxiv.org/abs/1506.01186,
sec. 3.3. The cycle lengths (step_size_up, step_size_down) are just
initial estimates and still need to be tuned.

The weights of CombinedLoss have been changed to [0.5, 0.5] to keep the
loss values in a similar range as its two components.

Closes #32