unet: Add experimental autoencoder mode
Using ae=True, the U-Net uses an alternative forward pass which doesn't
allow the decoder to access intermediate outputs of the encoder stages
(effectively removing long-range skip connections), thereby forcing the
decoder to only use the final output of the encoder to decode
features.
This is necessary for meaningful autoencoder training because
otherwise the network could just learn an identity mapping between the
input and the last decoder stage.
Using ae=True also has the effect that the decicated conv_ae_final
layer is used instead of the usual conv_final layer to deliver the final
output, so the same network model can easily be pretrained as an
autoencoder and later be switched to a segmentation network by setting
ae=False again.