training: Use torch's own DataLoader, fix #39
The private DataLoaderIter interface of PyTorch has changed again and
this time it's not as easy as renaming imports to fix the e3-internal
DelayedDataLoader class. Since the iterator class is now split into 2
parts and the implementation is selected at runtime, it has become
difficult to keep supporting the DelayedDataLoader mechanism, so I've
decided to just drop it (for now) and switch to raw
torch.utils.data.DataLoader.

This can lead to strange but harmless error messages
when resuming trainings after keyboard interrupts,
but I haven't encountered any actual issues in a batch of
test runs with this change.

If the user doesn't use keyboard interrupts to pause/resume the
training, this commit doesn't affect e3's behavior at all.