Revert unet ae mode (9b5c38d and 49d81ce)
This reverts commit the commits that introduced the experimental
autoencoder pretraining mode of the UNet model. It hasn't been used for
quite a while and introduced backwards compatibility issues when loading
pt models.
For now I'll just revert it completely. It could be re-introduced later in a
less breaky way or maybe a better idea would be to instead try the
approach of https://arxiv.org/abs/1810.11654 to better handle limited
training data scenarios.