example/unet: Disable deprecate adaptive conv
Adaptive convolution operators were intended to fix the very low
performance of 3D convolution operations with (1, 3, 3) shape.
As of PyTorch 1.2, CUDA 10.1 and CuDNN 7.6.2 this performance bug seems
to be fixed (not sure what fixed it, but it's no longer slow now).
Therefore adaptive=True shouldn't be necessary any more.

Another reason to disable it is that in PyTorch 1.2, training JIT-traced
models that use the 2D code path of the adaptive conv code fail due to outputs
not being contiguous.

Example trainings with --jit=train is now supported again.