warping: WIP elastic deformation integration
This code is not yet active because config options are missing,
but if enabled, this performs elastic deformations on the coordinate
level, *after* affine/perspective warping but *before* interpolating
and reading data.

The main advantage of preferring this over
elektronn3.data.transforms.ElasticTransform (which is more versatile and
works with 2D data as well) is that the costly
interpolation of float coordinates to the discrete voxel grid is then
only performed once, so it runs faster and more accurately while
at the same time needing less checks and special case handling because
that's already in the warping code.

scipy's gaussian_filter is still very slow (can take half a second for
one set of 3D coordinates), but this can surely be optimized, perhaps
with a PyTorch implementation or with numba.