unet: Support (almost) all input shapes
Previously, there was only support for input shapes that lead to
each DonwConv block's results being divisible in all dimensions in which
pooling was applied. This was especially making it tricky to find
compatible input shapes for conv_mode='valid' configurations.

This constraint is now removed because nn.MaxPool is now configured with
ceil_mode=True (output shapes from oddly-shaped inputs are calculated
with ceil, not floor) and the autocrop mechanism for merging tensors in
long-range skip connection now detects if encoder features need to be
cropped to match the upsampled decoder features and performs an
additional crop if necessary.