Add resunet model (U-Net with residual blocks)
Based on unet.py, but incompatible with weights from the old
implementation, so for compatibility reasons it is added as a separate
file. By default it works the same as the old U-Net, the only
differences are the internal implementation and the new options:
enc_res_blocks and dec_res_blocks.

This commit makes the "resunet" git branch obsolete.